# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BimbelFarmasi API Documentation",
        "description": "API Documentation untuk platform Bimbel Farmasi. Platform bimbingan belajar online untuk persiapan UKOM, CPNS & P3K Kefarmasian.",
        "contact": {
            "name": "BimbelFarmasi Support",
            "email": "bimbelfarmasi@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.bimbelfarmasi.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/programs/{slug}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get program by slug",
                "operationId": "fa094e41bc9412d2191673c7dbd7bcb7",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program not found"
                    }
                }
            }
        },
        "/api/v1/programs/{id}/materials": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get program materials",
                "operationId": "f926b560fbb304ab0442000baabf37f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program materials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/programs/{id}/exercises": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get program exercises",
                "operationId": "3d23b34cf7527111d8bf6c68e905c080",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program exercises"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/exercises/{exerciseId}/submit": {
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Submit exercise answers",
                "operationId": "0e838931b516b2fd57d85086648f1f97",
                "parameters": [
                    {
                        "name": "exerciseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "object",
                                        "example": {
                                            "1": "A",
                                            "2": "B",
                                            "3": "C"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answers submitted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}